T359 Main: Recognize output format even if parser fails
Use a simpler parser for second pass of parsing of arguments.
The simpler parser ignores unrecognized arguments and so has a better
chance of extracting the desired output format.
If JSON output is recognized as requested, the error is only printed
in JSON format.
Only the development version of Argparse4j currently supports ignoring
unrecognized arguments. The development version needs to be installed
when the project is set up. The instructions have been added in README.md.