OutPdfArgs: Expose `append` in a CL argument --append
The boolean command line option --append defaults to true.
If it is true, the original data in the PDF document will be kept intact
and the update will be recorded in a new revision.
If it is false, the PDF is "recompiled", possibly resulting in a smaller
document.
Previously append was always enabled in signatures and disabled in all
the other modifying operations.
Appending does not work with setting the version, i.e. to set PDF version,
--append must be set to false.