T473 `signature add`: Add support for timestamping
If --tsa-url is set, PDFMU tries to get a timestamp from the URL.
If --tsa-username is set, PDFMU tries to authorize the access to the TSA.
Authorization also uses a password set in --tsa-password or the envvar
pointed to by --tsa-password-envvar.
New error types have been introduced for the TSA errors:
- SIGNATURE_ADD_TSA_UNTRUSTED
- SIGNATURE_ADD_TSA_UNAUTHORIZED
- SIGNATURE_ADD_TSA_LOGIN_FAIL
- SIGNATURE_ADD_TSA_TRUSTSTORE_EMPTY

The TSA must found in the URL must be trusted by the TrustStore.
The default TrustStore may be overriden by setting for example
`-Djavax.net.ssl.trustStore="C:/Users/Filip/git/pdf-manipulation-utility/pdfmu/cacerts.jks"`
(you may pass this to `java`).