FIX: Global: Fix line endings to LF in all files
I use the NetBeans plugin License Changer to generate the license header
in the source files. The plugin always uses CRLF line endings
in the license (which seems to be a bug). I used  the bash command
`find -type f -exec dos2unix {} \;` in "pdfmu/src/main/java"
to fix the line endings to LF globally.